Use official API, fix image masks
The Automatic1111/stable-diffusion-webui API is now full-featured enough
that it can be used for almost all image generation operations, that's
now used instead of the /api/predict endpoint that's only intended for
internal gradio use.

Image masks no longer work as-is, so convert them to black and white RGB
to fix that problem.