Better fix for threading issues on single core machines
The problem wasn't duplicate signals, it was issues with the async
progress checks.  When only one thread is available in the thread pool,
the async progress check wouldn't start until the generation thread was
already finished.  It would then start, never stopping because it
wouldn't detect any progress, and it would block any new generate thread
from running.

There are two fixes for this: First, if the max thread pool count is
less than three, it will now be increased to three. Single-core systems
shouldn't have any problem with that, they just won't get any
performance boosts.  Second, SDWebUIGenerator will now assign an ID to
async tasks, so the async progress task knows to exit when that ID
changes in the generator.